mount.go split into multiple files
